requirements.txt
feedparser
pyairtable
schedule
python-dotenv
